+++ tests/planner/CMakeLists.txt
@@ -1,7 +1,7 @@
 add_subdirectory("parabolic")
 add_subdirectory("ompl")
 add_subdirectory("vectorfield")
-add_subdirectory("kinodynamic")
+add_subdirectory("optimalretimer")
 
 aikido_add_test(test_SnapPlanner test_SnapPlanner.cpp)
 target_link_libraries(test_SnapPlanner
